Refactor repository structure and add unit tests
