+++ app/src/main/java/com/example/getfast/repository/EbayRepository.kt
@@ -2,21 +2,26 @@ package com.example.getfast.repository
 
 import com.example.getfast.model.Listing
 import com.example.getfast.model.SearchFilter
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
-import org.jsoup.Jsoup
-import org.jsoup.nodes.Element
 
-class EbayRepository {
-    suspend fun fetchLatestListings(filter: SearchFilter): List<Listing> = withContext(Dispatchers.IO) {
+/**
+ * Repository zum Abrufen und Filtern von Wohnungsanzeigen.
+ * Die konkreten Schritte wie Netzwerkzugriff und Parsing
+ * werden über eigene Klassen delegiert, damit die Klasse
+ * klein und erweiterbar bleibt.
+ */
+class EbayRepository(
+    private val fetcher: HtmlFetcher = JsoupHtmlFetcher(),
+    private val parser: ListingParser = ListingParser(),
+) {
+    /**
+     * Lädt aktuelle Listings entsprechend dem Filter.
+     * Fehler führen zu einer leeren Liste.
+     */
+    suspend fun fetchLatestListings(filter: SearchFilter): List<Listing> {
         val url = "https://www.kleinanzeigen.de/s-wohnung-mieten/${filter.city.urlPath}"
-        try {
-            val doc = Jsoup.connect(url)
-                .userAgent("Mozilla/5.0")
-                .get()
-            val listings = doc.select("article.aditem").mapNotNull { element ->
-                parseListing(element)
-            }
+        return try {
+            val doc = fetcher.fetch(url)
+            val listings = parser.parse(doc)
             val maxPrice = filter.maxPrice
             if (maxPrice != null) {
                 listings.filter { listing ->
@@ -29,50 +34,6 @@ class EbayRepository {
             }
         } catch (e: Exception) {
             emptyList()
-        }
-    }
-
-    private fun parseListing(element: Element): Listing? {
-        val id = element.attr("data-adid")
-        val link = element.selectFirst("a[href].ellipsis")
-        val href = link?.attr("href")
-        val title = link?.text()
-        if (id.isEmpty() || href == null || title == null) {
-            return null
-        }
-        val date = element.selectFirst(".aditem-main--top--right")?.text()?.trim() ?: ""
-        val locationText = element.selectFirst(".aditem-main--top--left")?.text()?.trim() ?: ""
-        val parts = locationText.split(",")
-        val district = parts.getOrNull(0)?.trim() ?: ""
-        val city = parts.getOrNull(1)?.trim() ?: ""
-        val price = element.selectFirst(".aditem-main--middle--price-shipping")?.text()?.trim() ?: ""
-        val description = element.selectFirst(".aditem-main--middle--description")?.text()?.trim() ?: ""
-        val summary = generateSummary(description)
-        val images = element.select("img").mapNotNull {
-            it.attr("data-src").takeIf { src -> src.isNotBlank() }
-                ?: it.attr("src").takeIf { src -> src.isNotBlank() }
-        }.distinct()
-        val isSearch = listOf(title, description).any {
-            it.contains("suche", ignoreCase = true) || it.contains("gesuch", ignoreCase = true)
-        }
-        return Listing(
-            id = id,
-            title = title,
-            url = "https://www.kleinanzeigen.de$href",
-            date = date,
-            district = district,
-            city = city,
-            price = price,
-            summary = summary,
-            imageUrls = images,
-            isSearch = isSearch,
-        )
-    }
-
-    private fun generateSummary(text: String): String {
-        val sentences = text.split(".").map { it.trim() }.filter { it.isNotEmpty() }
-        return sentences.take(2).joinToString(". ").let {
-            if (it.isNotEmpty()) "$it." else ""
         }
     }
 }
